Add units. Refactor code.
