+++ react-components/src/__tests__/apps.test.js
@@ -3,11 +3,12 @@ import ReactDOM from 'react-dom';
 
 import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
 
-import { getCyVerseTheme } from './lib';
+import { getCyVerseTheme } from '../lib';
 
 import CategoryTreeTest from '../stories/apps/details/CategoryTree.stories';
 import ToolDetailsTest from '../stories/apps/details/ToolDetails.stories';
 import CopyTextAreaTest from '../stories/util/CopyTextArea.stories';
+import AppStatsTest from '../../stories/apps/admin/AppStats.stories';
 
 it('renders CopyTextArea without crashing', () => {
   const div = document.createElement('div');
@@ -26,3 +27,10 @@ it('renders ToolDetails without crashing', () => {
     ReactDOM.render(<MuiThemeProvider muiTheme={getCyVerseTheme()}><ToolDetailsTest /></MuiThemeProvider>, div);
     ReactDOM.unmountComponentAtNode(div);
 });
+
+it('renders ToolDetails without crashing', () => {
+    const div = document.createElement('div');
+    ReactDOM.render(<MuiThemeProvider muiTheme={getCyVerseTheme()}><AppStatsTest /></MuiThemeProvider>, div);
+    ReactDOM.unmountComponentAtNode(div);
+});
+
