+++ react-components/stories/data/details/Tag.stories.js
